refactor: move NavBar component in accordance with atomic design
