+++ src/pages/users/[id].tsx
@@ -1,6 +1,6 @@
 import { useRouter } from 'next/router';
 import Container from '../../components/atoms/Container';
-import NavBar from '../../components/NavBar';
+import NavBar from '../../components/molecules/NavBar';
 import TextList from '../../components/TextList';
 import UserAvatar from '../../components/UserAvatar';
 import UserName from '../../components/UserName';
